requirements.txt
streamlit
langgraph
langchain-google-genai
requests
python-dotenv